net: phy: mdio-gpio: Add platform_data support for phy_mask
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-193.el8
Rebuild_CHGLOG: - [kernel] phy: mdio-gpio: Add platform_data support for phy_mask (Petr Oros) [1772010]
Rebuild_FUZZ: 95.58%
commit-author Andrew Lunn <andrew@lunn.ch>
commit 04fa26bab06d9335f15a5d529c4bba25cd507a34

It is sometimes necessary to instantiate a bit-banging MDIO bus as a
platform device, without the aid of device tree.

When device tree is being used, the bus is not scanned for devices,
only those devices which are in device tree are probed. Without device
tree, by default, all addresses on the bus are scanned. This may then
find a device which is not a PHY, e.g. a switch. And the switch may
have registers containing values which look like a PHY. So during the
scan, a PHY device is wrongly created.

After the bus has been registered, a search is made for
mdio_board_info structures which indicates devices on the bus, and the
driver which should be used for them. This is typically used to
instantiate Ethernet switches from platform drivers.  However, if the
scanning of the bus has created a PHY device at the same location as
indicated into the board info for a switch, the switch device is not
created, since the address is already busy.

This can be avoided by setting the phy_mask of the mdio bus. This mask
prevents addresses on the bus being scanned.

v2
	Reviewed-by: Florian Fainelli <f.fainelli@gmail.com>

	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 04fa26bab06d9335f15a5d529c4bba25cd507a34)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>