iommu/arm-smmu-v3: Link domains and devices
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-193.el8
Rebuild_CHGLOG: - [iommu] arm-smmu-v3: Link domains and devices (Jerry Snitselaar) [1722978]
Rebuild_FUZZ: 92.50%
commit-author Jean-Philippe Brucker <jean-philippe.brucker@arm.com>
commit 2a7e62f51696ad476b7fb8157d73307e20257df7

When removing a mapping from a domain, we need to send an invalidation to
all devices that might have stored it in their Address Translation Cache
(ATC). In addition when updating the context descriptor of a live domain,
we'll need to send invalidations for all devices attached to it.

Maintain a list of devices in each domain, protected by a spinlock. It is
updated every time we attach or detach devices to and from domains.

It needs to be a spinlock because we'll invalidate ATC entries from
within hardirq-safe contexts, but it may be possible to relax the read
side with RCU later.

	Signed-off-by: Jean-Philippe Brucker <jean-philippe.brucker@arm.com>
	Signed-off-by: Will Deacon <will.deacon@arm.com>
(cherry picked from commit 2a7e62f51696ad476b7fb8157d73307e20257df7)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>