bus_find_device: Unify the match callback with class_find_device
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-193.el8
commit-author Suzuki K Poulose <suzuki.poulose@arm.com>
commit 418e3ea157efb0eb2c6dd412a8d5f052477c7f5a
Empty-Commit: Cherry-Pick Conflicts during history rebuild.
Will be included in final tarball splat. Ref for failed cherry-pick at:
ciq/ciq_backports/kernel-4.18.0-193.el8/418e3ea1.failed

There is an arbitrary difference between the prototypes of
bus_find_device() and class_find_device() preventing their callers
from passing the same pair of data and match() arguments to both of
them, which is the const qualifier used in the prototype of
class_find_device().  If that qualifier is also used in the
bus_find_device() prototype, it will be possible to pass the same
match() callback function to both bus_find_device() and
class_find_device(), which will allow some optimizations to be made in
order to avoid code duplication going forward.  Also with that, constify
the "data" parameter as it is passed as a const to the match function.

For this reason, change the prototype of bus_find_device() to match
the prototype of class_find_device() and adjust its callers to use the
const qualifier in accordance with the new prototype of it.

	Cc: Alexander Shishkin <alexander.shishkin@linux.intel.com>
	Cc: Andrew Lunn <andrew@lunn.ch>
	Cc: Andreas Noever <andreas.noever@gmail.com>
	Cc: Arnd Bergmann <arnd@arndb.de>
	Cc: Bjorn Helgaas <bhelgaas@google.com>
	Cc: Corey Minyard <minyard@acm.org>
	Cc: Christian Borntraeger <borntraeger@de.ibm.com>
	Cc: David Kershner <david.kershner@unisys.com>
	Cc: "David S. Miller" <davem@davemloft.net>
	Cc: David Airlie <airlied@linux.ie>
	Cc: Felipe Balbi <balbi@kernel.org>
	Cc: Frank Rowand <frowand.list@gmail.com>
	Cc: Grygorii Strashko <grygorii.strashko@ti.com>
	Cc: Harald Freudenberger <freude@linux.ibm.com>
	Cc: Hartmut Knaack <knaack.h@gmx.de>
	Cc: Heiko Stuebner <heiko@sntech.de>
	Cc: Jason Gunthorpe <jgg@ziepe.ca>
	Cc: Jonathan Cameron <jic23@kernel.org>
	Cc: "James E.J. Bottomley" <jejb@linux.ibm.com>
	Cc: Len Brown <lenb@kernel.org>
	Cc: Mark Brown <broonie@kernel.org>
	Cc: Michael Ellerman <mpe@ellerman.id.au>
	Cc: Michael Jamet <michael.jamet@intel.com>
	Cc: "Martin K. Petersen" <martin.petersen@oracle.com>
	Cc: Peter Oberparleiter <oberpar@linux.ibm.com>
	Cc: Sebastian Ott <sebott@linux.ibm.com>
	Cc: Srinivas Kandagatla <srinivas.kandagatla@linaro.org>
	Cc: Yehezkel Bernat <YehezkelShB@gmail.com>
	Cc: rafael@kernel.org
	Acked-by: Corey Minyard <minyard@acm.org>
	Acked-by: David Kershner <david.kershner@unisys.com>
	Acked-by: Mark Brown <broonie@kernel.org>
	Acked-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
	Acked-by: Srinivas Kandagatla <srinivas.kandagatla@linaro.org>
	Acked-by: Wolfram Sang <wsa@the-dreams.de> # for the I2C parts
	Acked-by: Rob Herring <robh@kernel.org>
	Signed-off-by: Suzuki K Poulose <suzuki.poulose@arm.com>
	Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 418e3ea157efb0eb2c6dd412a8d5f052477c7f5a)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>

# Conflicts:
#	drivers/acpi/sleep.c
#	drivers/base/devcon.c
#	drivers/hwtracing/coresight/coresight.c
#	drivers/i2c/i2c-core-of.c