signal/bpfilter: Fix bpfilter_kernl to use send_sig not force_sig
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-193.el8
commit-author Eric W. Biederman <ebiederm@xmission.com>
commit 1dfd1711de2952fd1bfeea7152bd1687a4eea771

The locking in force_sig_info is not prepared to deal with
a task that exits or execs (as sighand may change).  As force_sig
is only built to handle synchronous exceptions.

Further the function force_sig_info changes the signal state if the
signal is ignored, or blocked or if SIGNAL_UNKILLABLE will prevent the
delivery of the signal.  The signal SIGKILL can not be ignored and can
not be blocked and SIGNAL_UNKILLABLE won't prevent it from being
delivered.

So using force_sig rather than send_sig for SIGKILL is pointless.

Because it won't impact the sending of the signal and and because
using force_sig is wrong, replace force_sig with send_sig.

	Cc: Alexei Starovoitov <ast@kernel.org>
	Cc: David S. Miller <davem@davemloft.net>
Fixes: d2ba09c17a06 ("net: add skeleton of bpfilter kernel module")
	Signed-off-by: "Eric W. Biederman" <ebiederm@xmission.com>
(cherry picked from commit 1dfd1711de2952fd1bfeea7152bd1687a4eea771)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>