net/mlx5: E-Switch, Fix access to invalid memory when toggling esw modes
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-193.el8
Rebuild_CHGLOG: - [netdrv] mlx5: E-Switch, Fix access to invalid memory when toggling esw modes (Alaa Hleihel) [1722659 1724333]
Rebuild_FUZZ: 97.14%
commit-author Roi Dayan <roid@mellanox.com>
commit 8a91ad9355c66c5026d3d911b434a25408ab876c

The esw fdb table has a union of legacy and offloads members.
So if we were in a certain esw mode we could set some memebers and not
set null which is fine as on destroy path and don't care.
But then moving from legacy to switchdev a second time, the cleanup flow
of legacy mode checks if a struct member was in use if it's not null so
we need to make sure to reset the code to null when we init legacy mode.

Fixes: 8da202b24913 ("net/mlx5: E-Switch, Add support for VEPA in legacy mode.")
	Signed-off-by: Roi Dayan <roid@mellanox.com>
	Reviewed-by: Huy Nguyen <huyn@mellanox.com>
	Signed-off-by: Saeed Mahameed <saeedm@mellanox.com>
(cherry picked from commit 8a91ad9355c66c5026d3d911b434a25408ab876c)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>