selftests: forwarding: test for bridge mcast traffic after report and leave
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-193.el8
commit-author Nikolay Aleksandrov <nikolay@cumulusnetworks.com>
commit f1054c65bca637c220fe8e32648156459361bb99

This test is split in two, the first part checks if a report creates a
corresponding mdb entry and if traffic is properly forwarded to it, and
the second part checks if the mdb entry is deleted after a leave and
if traffic is *not* forwarded to it. Since the mcast querier is enabled
we should see standard mcast snooping bridge behaviour.

	Signed-off-by: Nikolay Aleksandrov <nikolay@cumulusnetworks.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit f1054c65bca637c220fe8e32648156459361bb99)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>