mlxsw: spectrum_buffers: Use new CPU ingress pool for control packets
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-193.el8
commit-author Ido Schimmel <idosch@mellanox.com>
commit 50b5b90514a8306c4a8e33f967f15bac5c9ac1f5

Use the new ingress pool that was added in the previous patch for
control packets (e.g., STP, LACP) that are trapped to the CPU.

The previous management pool is no longer necessary and therefore its
size is set to 0.

The maximum quota for traffic towards the CPU is increased to 50% of the
free space in the new ingress pool and therefore the reserved space is
reduced by half, to 10KB - in both the shared and headroom buffer. This
allows for more efficient utilization of the shared buffer as reserved
space cannot be used for other purposes.

	Signed-off-by: Ido Schimmel <idosch@mellanox.com>
	Acked-by: Jiri Pirko <jiri@mellanox.com>
	Reviewed-by: Petr Machata <petrm@mellanox.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 50b5b90514a8306c4a8e33f967f15bac5c9ac1f5)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>