selinux: always return a secid from the network caches if we find one
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-193.el8
commit-author Paul Moore <paul@paul-moore.com>
commit 9b80c36353ed4cce324af21244a65984db21991b

Previously if we couldn't find an entry in the cache and we failed to
allocate memory for a new cache entry we would fail the network object
label lookup; this is obviously not ideal.  This patch fixes this so
that we return the object label even if we can't cache the object at
this point in time due to memory pressure.

The GitHub issue tracker is below:
 * https://github.com/SELinuxProject/selinux-kernel/issues/3

	Signed-off-by: Paul Moore <paul@paul-moore.com>
(cherry picked from commit 9b80c36353ed4cce324af21244a65984db21991b)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>