KVM: arm64: pmu: Fix cycle counter truncation
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-193.el8
commit-author Marc Zyngier <maz@kernel.org>
commit f4e23cf9477452640d2522f3286db5df6fcfe45a

When a counter is disabled, its value is sampled before the event
is being disabled, and the value written back in the shadow register.

In that process, the value gets truncated to 32bit, which is adequate
for any counter but the cycle counter (defined as a 64bit counter).

This obviously results in a corrupted counter, and things like
"perf record -e cycles" not working at all when run in a guest...
A similar, but less critical bug exists in kvm_pmu_get_counter_value.

Make the truncation conditional on the counter not being the cycle
counter, which results in a minor code reorganisation.

Fixes: 80f393a23be6 ("KVM: arm/arm64: Support chained PMU counters")
	Reviewed-by: Andrew Murray <andrew.murray@arm.com>
	Reported-by: Julien Thierry <julien.thierry.kdev@gmail.com>
	Signed-off-by: Marc Zyngier <maz@kernel.org>
(cherry picked from commit f4e23cf9477452640d2522f3286db5df6fcfe45a)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>