powerpc/eeh: Fix race when freeing PDNs
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-193.el8
Rebuild_CHGLOG: - [powerpc] eeh: Fix race when freeing PDNs (Gustavo Duarte) [1524642]
Rebuild_FUZZ: 88.57%
commit-author Oliver O'Halloran <oohall@gmail.com>
commit 5ef753ae435a5cea8af5c84a65fc5dd30b773040

When hot-adding devices we rely on the hotplug driver to create pci_dn's
for the devices under the hotplug slot. Converse, when hot-removing the
driver will remove the pci_dn's that it created. This is a problem because
the pci_dev is still live until it's refcount drops to zero. This can
happen if the driver is slow to tear down it's internal state. Ideally, the
driver would not attempt to perform any config accesses to the device once
it's been marked as removed, but sometimes it happens. As a result, we
might attempt to access the pci_dn for a device that has been torn down and
the kernel may crash as a result.

To fix this, don't free the pci_dn unless the corresponding pci_dev has
been released.  If the pci_dev is still live, then we mark the pci_dn with
a flag that indicates the pci_dev's release function should free it.

	Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
	Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Link: https://lore.kernel.org/r/20190903101605.2890-3-oohall@gmail.com
(cherry picked from commit 5ef753ae435a5cea8af5c84a65fc5dd30b773040)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>