dm writecache: avoid unnecessary lookups in writecache_find_entry()
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-193.el8
commit-author Mikulas Patocka <mpatocka@redhat.com>
commit f8011d334426cee77276a1038b627b5cb0470258

This is a small optimization in writecache_find_entry().

If we go past the condition "if (unlikely(!node))", we can be certain that
there is no entry in the tree that has the block equal to the "block"
variable.

Consequently, we can return the next entry directly, we don't need to go
to the second part of the function that finds the entry with lowest or
highest seq number that matches the "block" variable.

Also, add some whitespace and cleanup needless braces.

	Suggested-by: Huaisheng Ye <yehs1@lenovo.com>
	Signed-off-by: Mikulas Patocka <mpatocka@redhat.com>
	Signed-off-by: Mike Snitzer <snitzer@redhat.com>
(cherry picked from commit f8011d334426cee77276a1038b627b5cb0470258)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>