dm stripe: use struct_size() in kmalloc()
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-193.el8
commit-author Gustavo A. R. Silva <gustavo@embeddedor.com>
commit 8adeac3be03d400f9c2391d52f85cd27bd188800

One of the more common cases of allocation size calculations is finding
the size of a structure that has a zero-sized array at the end, along
with memory for some number of elements for that array. For example:

struct stripe_c {
        ...
        struct stripe stripe[0];
};

In this case alloc_context() and dm_array_too_big() are removed and
replaced by the direct use of the struct_size() helper in kmalloc().

Notice that open-coded form is prone to type mistakes.

This code was detected with the help of Coccinelle.

	Signed-off-by: Gustavo A. R. Silva <gustavo@embeddedor.com>
	Signed-off-by: Mike Snitzer <snitzer@redhat.com>
(cherry picked from commit 8adeac3be03d400f9c2391d52f85cd27bd188800)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>