jump_label: Sort entries of the same key by the code
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-193.el8
commit-author Daniel Bristot de Oliveira <bristot@redhat.com>
commit 0f133021bd82548a33580bfb7b055e8857f46c2a

In the batching mode, all the entries of a given key are updated at once.
During the update of a key, a hit in the int3 handler will check if the
hitting code address belongs to one of these keys.

To optimize the search of a given code in the vector of entries being
updated, a binary search is used. The binary search relies on the order
of the entries of a key by its code. Hence the keys need to be sorted
by the code too, so sort the entries of a given key by the code.

	Signed-off-by: Daniel Bristot de Oliveira <bristot@redhat.com>
	Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
	Cc: Borislav Petkov <bp@alien8.de>
	Cc: Chris von Recklinghausen <crecklin@redhat.com>
	Cc: Clark Williams <williams@redhat.com>
	Cc: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
	Cc: H. Peter Anvin <hpa@zytor.com>
	Cc: Jason Baron <jbaron@akamai.com>
	Cc: Jiri Kosina <jkosina@suse.cz>
	Cc: Josh Poimboeuf <jpoimboe@redhat.com>
	Cc: Linus Torvalds <torvalds@linux-foundation.org>
	Cc: Marcelo Tosatti <mtosatti@redhat.com>
	Cc: Masami Hiramatsu <mhiramat@kernel.org>
	Cc: Peter Zijlstra <peterz@infradead.org>
	Cc: Scott Wood <swood@redhat.com>
	Cc: Steven Rostedt (VMware) <rostedt@goodmis.org>
	Cc: Thomas Gleixner <tglx@linutronix.de>
Link: https://lkml.kernel.org/r/f57ae83e0592418ba269866bb7ade570fc8632e0.1560325897.git.bristot@redhat.com
	Signed-off-by: Ingo Molnar <mingo@kernel.org>
(cherry picked from commit 0f133021bd82548a33580bfb7b055e8857f46c2a)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>