sysfs: Fix internal_create_group() for named group updates
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-193.el8
commit-author Rajat Jain <rajatja@google.com>
commit c855cf2759d27142f771173d9fd8e7fdf9cf5138

There are a couple of problems with named group updates in the code
today:

* sysfs_update_group() will always fail for a named group, because
  internal_create_group() will try to create a new sysfs directory
  unconditionally, which will ofcourse fail with -EEXIST.

* We can leak the kernfs_node for grp->name if some one tries to:
  - rename a group (change grp->name), or
  - update a named group, to an unnamed group

It appears that the whole purpose of sysfs_update_group() was to
allow changing the permissions or visibility of attributes and not
the names. So make it clear in the comments, and allow it to update
an existing named group.

	Signed-off-by: Rajat Jain <rajatja@google.com>
	Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit c855cf2759d27142f771173d9fd8e7fdf9cf5138)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>