perf db-export: Pass main_thread to db_export__thread()
jira LE-1907
Rebuild_History Non-Buildable kernel-4.18.0-193.el8
commit-author Adrian Hunter <adrian.hunter@intel.com>
commit ed5c0a16feb9f1a4347f109d5e9607f6f38688a0

Calls to db_export__thread() already have main_thread so there is no
reason to get it again, instead pass it as a parameter. Note that one
difference in this approach is that the main thread is not created if it
does not exist. It is better if it is not created because:

   - If main_thread is being traced it will have been created already.

   - If it is not being traced, there will be no other information about
     it, and it will never get deleted because there will be no EXIT event.

	Signed-off-by: Adrian Hunter <adrian.hunter@intel.com>
	Cc: Jiri Olsa <jolsa@redhat.com>
Link: http://lkml.kernel.org/r/20190710085810.1650-4-adrian.hunter@intel.com
	Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>
(cherry picked from commit ed5c0a16feb9f1a4347f109d5e9607f6f38688a0)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>